backend/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
sqlalchemy
psycopg2-binary
python-dotenv
pandas
pydantic
pydantic-settings
jinja2
weasyprint
